src-account/requirements.txt
uritemplate.py==3.0.2
asgiref==3.5.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.6
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.2
django-toolbelt==0.0.1
djangorestframework==3.13.1
gunicorn==20.1.0
MarkupSafe==2.1.1
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytest==7.1.1
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.41
sqlparse==0.4.2
swagger-ui-bundle==0.0.9
python-dotenv==0.21.1
django-rest-auth==0.9.5
dj-rest-auth==2.2.7
django-allauth==0.52.0
drf-yasg==1.21.5
      

